+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added: Generator command for sample crontab.xml file
 - Added: Generator command for sample email_templates.xml file
 - Added: Generator command for sample sections.xml file
+- Added: Generator command for sample fieldset.xml file
 
 ## [1.1.3] - 3/12/2025